Add assert statements to Storage and TaskList to ensure correctness of assumptions at points of execution
The Storage class checks that the save file for tasks when it is initialised, creating it if it doesn't. From there, the file's existence is no longer checked for. To check that the assumption that the file will continue to exist throughout the program's runtime is true, let's add assert statements to each method of Storage.

The TaskList class contains a few methods that operates on a task, taking the index of the task as the parameter. Currently, the level of abstraction above the TaskList, which is the Command classes, do checks that the index is valid, before calling these methods, thus, the TaskList class itself does not check the index's validity again. To check that these methods are never called with invalid indexes, let's add assert statements to each of these methods.