Modify Task to implement Comparable interface so that there is a natural order of tasks
Tasks in the tasklist are currently always displayed in the order that they are added. This may not be the most useful way to show tasks to the user.

One obvious ordering of tasks that will be of interest to the user is chronological. Let's implement a natural order within Task classes that uses their given LocalDateTime attribute as the point of comparison. In addition, the TaskList class will implement a sort method that will allow other ways of sorting in the future.